Some refactoring and cleanup, added new drivers etc.
Reorganized the HAL structures a bit for cleaner code, harmonized driver
options across all drivers, added driver code for SAMD21 and
MSP432E401Y, updated TMC129x code to FreeRTOS version with ethernet and
SD card options++

+++ drivers/TM4C123/keypad.c
@@ -221,7 +221,7 @@ static char *strrepl (char *str, int c, char *str3)
 	return str;
 }
 
-void process_keypress (uint8_t state)
+void process_keypress (uint_fast16_t state)
 {
 	bool addedGcode, jogCommand = false;
 	char command[30] = "", keycode = keypad_get_keycode();
